requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

boto3==1.35.10
    # via -r requirements.in
botocore==1.35.10
    # via
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deprecated==1.2.14
    # via pikepdf
exceptiongroup==1.2.2
    # via pytest
fonttools==4.53.1
    # via matplotlib
fuzzywuzzy[speedup]==0.18.0
    # via -r requirements.in
idna==3.8
    # via requests
img2pdf==0.5.1
    # via pdf2jpg
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.4.2
    # via nltk
kiwisolver==1.4.5
    # via matplotlib
levenshtein==0.25.1
    # via python-levenshtein
lxml==5.3.0
    # via pikepdf
matplotlib==3.9.2
    # via -r requirements.in
nltk==3.9.1
    # via -r requirements.in
numpy==2.1.0
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
packaging==24.1
    # via
    #   matplotlib
    #   pikepdf
    #   pytest
pandas==2.2.2
    # via -r requirements.in
pdf2jpg==1.1
    # via -r requirements.in
pikepdf==9.2.1
    # via img2pdf
pillow==10.4.0
    # via
    #   -r requirements.in
    #   img2pdf
    #   matplotlib
    #   pikepdf
pluggy==1.5.0
    # via pytest
pyparsing==3.1.4
    # via matplotlib
pypdf2==3.0.1
    # via -r requirements.in
pytest==8.3.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
python-levenshtein==0.25.1
    # via
    #   -r requirements.in
    #   fuzzywuzzy
pytz==2024.1
    # via pandas
rapidfuzz==3.9.7
    # via
    #   -r requirements.in
    #   levenshtein
regex==2024.7.24
    # via nltk
requests==2.32.3
    # via -r requirements.in
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via pytest
tqdm==4.66.5
    # via nltk
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   botocore
    #   requests
wrapt==1.16.0
    # via deprecated
